requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langsmith
python-dotenv
langchain-groq
gradio
pandas
numpy
scikit-learn
torch
transformers>=4.44.0
accelerate
sentencepiece
peft
trl
sendgrid
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai
langchain_tavily>=0.1.0
langchain_huggingface>=0.1.0
fastapi
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client